feat(toolkit): promote runtime context package
