+++ .changeset/hook-context-inline.md
@@ -1,7 +1,5 @@
 ---
 "skillset": patch
-"@skillset/core": patch
-"@skillset/schema": patch
 ---
 
 Add inline adaptive hook runtime context delivery while reserving toolkit-backed hooks until helper provenance is ready.
