+++ .changeset/toolkit-runtime-package.md
@@ -1,0 +1,5 @@
+---
+"skillset": patch
+---
+
+Promote the runtime context helper into the private `@skillset/toolkit/runtime` workspace package so hook scripts can build against the same typed provider context model the CLI uses.
