+++ apps/skillset/src/changeset-awareness.ts
@@ -50,6 +50,7 @@ export function isPackageAffectingPath(path: string) {
   if (isRuntimeSourcePath(path, "packages/lint/src")) return true;
   if (isRuntimeSourcePath(path, "packages/provider-formats/src")) return true;
   if (isRuntimeSourcePath(path, "packages/schema/src")) return true;
+  if (isRuntimeSourcePath(path, "packages/toolkit/src")) return true;
   if (isRuntimeSourcePath(path, "packages/transforms/src")) return true;
 
   return (
@@ -57,6 +58,7 @@ export function isPackageAffectingPath(path: string) {
     path === "packages/lint/package.json" ||
     path === "packages/provider-formats/package.json" ||
     path === "packages/schema/package.json" ||
+    path === "packages/toolkit/package.json" ||
     path === "packages/transforms/package.json"
   );
 }
