+++ docs/features/hook-guardrails.md
@@ -31,7 +31,7 @@ skillset hooks context --event Stop --format env --context-fields provider,hook.
 | Git hook-runner snippet | n/a | n/a | `implemented` | Prints additive snippets for existing hook runners; does not install. |
 | Agent runtime hook suggestion | reviewed config suggestion | reviewed config suggestion | `implemented` / `target_specific` | Prints project-local suggestions; must not mutate runtime config automatically. |
 | Agent runtime hook execution | `skillset hooks run post-tool-use`, `skillset hooks run stop` | `skillset hooks run post-tool-use`, `skillset hooks run stop` | `implemented` / `target_specific` | Core CLI behavior called by reviewed project-local runtime config. |
-| Runtime context helper | `skillset hooks context --event <event> --format env|json` | `skillset hooks context --event <event> --format env|json` | `implemented` / `target_specific` | Internal helper used by generated adaptive hook wrappers for `context.strategy: toolkit`; public package surface is tracked separately. |
+| Runtime context helper | `skillset hooks context --event <event> --format env|json` | `skillset hooks context --event <event> --format env|json` | `implemented` / `target_specific` | Internal helper used by generated adaptive hook wrappers for `context.strategy: toolkit`; the shared context model lives in `@skillset/toolkit/runtime`. |
 
 ## Diagnostics
