+++ docs/features/hooks.md
@@ -86,7 +86,7 @@ Adaptive hook units can choose how much Skillset-normalized runtime context to p
 | --- | --- |
 | `inline` | Prefixes the generated command with requested `SKILLSET_*` environment assignments. |
 | `none` | Passes no Skillset-normalized context. Provider-native environment remains available. |
-| `toolkit` | Renders through Skillset's internal `skillset hooks context` helper. This proves the helper-backed wrapper path before the public `@skillset/toolkit/runtime` package boundary lands. |
+| `toolkit` | Renders through Skillset's internal `skillset hooks context` helper, backed by the shared `@skillset/toolkit/runtime` context model. Generated hooks still use the CLI helper until the compiler integration slice switches wrappers to the package/shell boundary. |
 
 Inline v1 fields are deliberately small:
 
@@ -104,7 +104,7 @@ For `context.strategy: toolkit`, generated hooks call:
 skillset hooks context --event <event> --format env --context-fields <field,...>
 ```
 
-The helper prints shell `export` statements for the requested `SKILLSET_*` values and preserves provider-native environment access for the hook command. This is an internal proof surface for generated wrappers; public JS/TS and shell package surfaces are tracked separately under the `@skillset/toolkit` follow-up stack.
+The helper prints shell `export` statements for the requested `SKILLSET_*` values and preserves provider-native environment access for the hook command. Hook scripts that run inside this repo can also import the typed package surface from `@skillset/toolkit/runtime`; generated hook output does not depend on that package until the compiler integration slice lands.
 
 ### Attachments
