+++ packages/toolkit/package.json
@@ -1,0 +1,10 @@
+{
+  "name": "@skillset/toolkit",
+  "version": "0.0.0",
+  "private": true,
+  "type": "module",
+  "exports": {
+    ".": "./src/index.ts",
+    "./runtime": "./src/runtime.ts"
+  }
+}
